requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.4
agate-sql==0.5.7
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
csvkit==1.0.6
dbfread==2.0.7
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
Faker==8.10.3
future==0.18.2
greenlet==1.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
leather==0.3.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mixer==7.1.2
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.21.1
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
parsedatetime==2.4
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-ipdb @ git+git://github.com/mverteuil/pytest-ipdb.git@8e7f57b94e41832795c7171c999828b2999bea91
python-dateutil==2.8.2
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.1
regex==2021.7.6
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.1.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
xlrd==2.0.1
